requirements.txt
ai2thor>=5.0.0
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
cmake==3.25.0
contourpy==1.0.5
cryptography==41.0.2
cycler==0.11.0
decorator==5.1.1
diskcache==5.6.1
einops==0.6.1
executing==1.2.0
filelock==3.12.2
Flask==1.1.2
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2023.6.0
gptcache==0.1.37
hydra-core==1.3.2
idna==3.4
importlib-resources==5.2.0
ipdb==0.13.13
ipython==8.12.2
itsdangerous==1.1.0
jedi==0.18.2
Jinja2==2.11.3
kiwisolver==1.4.4
lit==15.0.7
MarkupSafe==1.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.2.1
msal==1.23.0
multidict==6.0.4
munkres==1.1.4
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
omegaconf==2.3.0
openai==0.27.8
opencv-python==4.8.0.74
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.1.2
platformdirs==3.9.1
ply==3.11
progressbar2==4.2.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-utils==3.7.0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
revtok==0.0.3
scipy==1.10.1
setuptools==67.8.0
sip==6.6.2
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tiktoken==0.4.0
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.3
vocab==0.0.5
wcwidth==0.2.6
Werkzeug==1.0.1
wheel==0.38.4
yarl==1.9.2
zipp==3.11.0
